pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["Talafhamohammad-cloud <josephtelfah@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
